feat: move reminder logic into modular task
