+++ bot/utils/config.py
@@ -1,6 +1,8 @@
 import os
 import json
 
+from pymongo import MongoClient
+
 config = {}
 if os.environ.get("ENVIRONMENT") == "production":
   config = {
@@ -14,3 +16,13 @@ if os.environ.get("ENVIRONMENT") == "production":
 else:
   config_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "config.json")
   config = json.load(open(config_path))
+
+print("Initializing database")
+
+client = MongoClient(config["MONGODB_URI"])
+if os.environ.get("ENVIRONMENT") == "DEVELOPMENT":
+  database = client.appeal_dev
+  print("Set appeal database to development environment")
+else:
+  database = client.appeal
+  print("Set appeal database to production")
